requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
torchmetrics>=0.11.0
tensorboard>=2.10.0
pyyaml>=6.0
numpy>=1.21.0
matplotlib>=3.5.0
pillow>=9.0.0
scikit-learn>=1.1.0
tqdm>=4.64.0
